backend/requirements/base.txt
# Core
python-slugify==8.0.4
Pillow==11.1.0
argon2-cffi==23.1.0

# Redis
redis==5.2.1
hiredis==3.1.0

# Celery
celery==5.5.0
django-celery-beat==2.7.0

# Django
django==5.1.8
django-environ==0.12.0
django-model-utils==5.0.0
django-redis==5.4.0

# Django REST Framework
djangorestframework==3.16.0
django-cors-headers==4.7.0
drf-spectacular==0.28.0
django-filter==24.3

# Server
uvicorn==0.34.0
gunicorn==23.0.0
watchfiles==1.0.0

# Celery monitoring
flower==2.0.1

# Health checks
django-health-check==3.18.3

# JWT/Clerk Authentication
PyJWT==2.10.0
cryptography==44.0.0
httpx==0.28.1

# Database
psycopg[binary]==3.2.4

# Storage (S3-compatible)
django-storages[s3]==1.14.4
boto3==1.35.99
